feat: implement global precision handling and device consistency
- Add GLOBAL_PT_FLOAT_PRECISION and GLOBAL_PT_ENER_FLOAT_PRECISION in env.py
- Replace hardcoded float64 with dynamic precision based on input tensor types
- Fix device placement issues across all modules
- Improve error handling for missing dependencies
- Use torch.inverse instead of torch.linalg.inverse for compatibility
- Update buffer_scales to use correct device placement

+++ src/torch_admp/optimizer.py
@@ -50,7 +50,7 @@ def line_search(
     torch.Tensor
         Optimal point found by line search
     """
-    history_x = torch.arange(3, dtype=torch.float64, device=x0.device)
+    history_x = torch.arange(3, dtype=x0.dtype, device=x0.device)
     if fk is None:
         x0 = x0.detach()
         x0.requires_grad = True
